requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
python-dotenv==1.1.1
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
